Peaking algorithm working. A period estimation can now be calculated.

+++ src/plotter.cpp
@@ -17,7 +17,7 @@ void plotter::write1DPlotData(const char* filename, short * data, int data_size)
     fs.close();
 }
 
-void plotter::write1DPlotData(const char* filename, float * data, int data_size){
+void plotter::write1DPlotData(const char* filename, double * data, int data_size){
     fstream fs;
     fs.open (filename, fstream::out);
