+++ src/plotter.h
@@ -8,7 +8,7 @@ class plotter{
 public:
 plotter();
 void write1DPlotData(const char* filename, short * data, int data_size);
-void write1DPlotData(const char* filename, float * data, int data_size);
+void write1DPlotData(const char* filename, double * data, int data_size);
 
     
 };
